3.1.12 Reorganizacja projektu, usuniecie publicznej klasy Node
